image registration
Depth data is now resized to half the size of color resolution, while
color is downsampled to half its size. So called image registration. The
actual way of doing it is referring its UV value in UVMap and find the
corresponding value, but there's no such precise need as no point cloud
is being used here.

+++ cvtest/cvtest/ImgRegistration.h
@@ -7,6 +7,7 @@
 #include <fstream>
 #include <math.h>
 #include <time.h>
+
 #include <pcl/sample_consensus/ransac.h>
 #include <pcl/sample_consensus/sac_model_plane.h>
 #include <pcl/common/common_headers.h>
@@ -25,7 +26,7 @@ public :
 	double getCrossVal();
 	double ImageRegistration::calcCrossVal(cv::Mat img1, cv::Mat img2, int offx, int offy, cv::Size window);
 	std::pair<int,int> getOffset();
-	std::pair<std::pair<int,int>,double> Norm_CrossCorr(cv::Mat L_src, cv::Mat R_src, double startx, double starty, cv::Size window);
+	std::pair<std::pair<int,int>,double> Norm_CrossCorr(cv::Mat L_src, cv::Mat R_src, int startx, int starty, cv::Size window);
 	pcl::PointCloud<pcl::PointXYZ>::Ptr RANSAC(pcl::PointCloud<pcl::PointXYZ>::Ptr inputcloud);
 	pcl::PointCloud<pcl::PointXYZ>::Ptr cvtMat2Cloud(cv::Mat* src);
 	cv::Mat cvCloud2Mat(pcl::PointCloud<pcl::PointXYZ>::Ptr cloud);
